real_state/requirements.txt
Django==4.2.*
uvicorn==0.21.1
fastapi==0.95.1
python-dotenv~=0.21.0
psycopg2-binary==2.9.3
uWSGI~=2.0.20
gunicorn==20.1.0
django-crispy-forms==1.14.0
django-cors-headers==3.14.0
pymongo==4.3.3
mixer==7.2.2
pandas==2.0.0
pytest-django==4.5.2
httpx==0.24.0
openai==0.27.0
